+++ src/utilities/index.ts
@@ -1,3 +1,4 @@
 export * from './alert';
+export * from './ffmpeg';
 export * from './firebase';
 export * from './functions';